(fix): Reports webview should display parameter input based on parameter type
