+++ packages/esm-reports-app/src/components/reports-data-overview.component.tsx
@@ -1,15 +1,17 @@
 import React, { useState, useMemo, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Select, SelectItem, Button } from '@carbon/react';
-import { ExtensionSlot, showSnackbar, OpenmrsDatePicker, getCoreTranslation } from '@openmrs/esm-framework';
-import { useReportDefinitions, useReportData } from './reports.resource';
+import dayjs from 'dayjs';
+import { ExtensionSlot, showSnackbar, getCoreTranslation } from '@openmrs/esm-framework';
 import Overlay from './overlay.component';
 import ReportDataViewer from './report-data-viewer.component';
+import ReportParameterBasedElement from './report-parameter-based-element';
+import { useReportDefinitions, useReportData, useLocations } from './reports.resource';
 import styles from './reports.scss';
-import dayjs from 'dayjs';
 
 const ReportsDataOverviewComponent: React.FC = () => {
   const { t } = useTranslation();
+  const { locations } = useLocations();
   const [selectedReport, setSelectedReport] = useState('');
   const [reportParameters, setReportParameters] = useState<Record<string, string>>({});
 
@@ -20,13 +22,24 @@ const ReportsDataOverviewComponent: React.FC = () => {
     return reportDefinitions?.find((report) => report.uuid === selectedReport);
   }, [selectedReport, reportDefinitions]);
 
-  const handleParameterChange = (parameterName: string, value: Date) => {
+  const handleDateChange = (parameterName: string, value: Date) => {
     const formattedDate = value ? dayjs(value).format('YYYY-MM-DDTHH:mm:ss.SSSZ') : '';
     setReportParameters((prev) => ({
       ...prev,
       [parameterName]: formattedDate,
     }));
   };
+  function handleParameterValueChange(event) {
+    const key = event.target.name;
+    let value = null;
+    if (event.target.type === 'checkbox') {
+      value = event.target.checked;
+    } else {
+      value = event.target.value;
+    }
+
+    setReportParameters((state) => ({ ...state, [key]: value }));
+  }
 
   const handleReportChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
     const newReportUuid = event.target.value;
@@ -116,11 +129,13 @@ const ReportsDataOverviewComponent: React.FC = () => {
           </div>
           {selectedReportDefinition?.parameters?.map((parameter) => (
             <div key={parameter.name + selectedReportDefinition.uuid} className={styles.formField}>
-              <OpenmrsDatePicker
-                id={parameter.name}
-                labelText={parameter.label}
-                onChange={(date) => handleParameterChange(parameter.name, date)}
-                value={reportParameters[parameter.name]}
+              <ReportParameterBasedElement
+                parameter={parameter}
+                reportUuid={selectedReport}
+                reportParameters={reportParameters}
+                locations={locations}
+                handleOnChange={handleParameterValueChange}
+                handleOnDateChange={handleDateChange}
               />
             </div>
           ))}
