+++ packages/esm-reports-app/src/components/run-report/run-report-form.scss
@@ -60,11 +60,6 @@
   padding-left: 0px;
 }
 
-.unknownParameterTypeSpan {
-  text-align: left;
-  color: red;
-}
-
 .tableIcon {
   display: inline-block;
   margin-left: 10px;
@@ -79,9 +74,9 @@
 }
 
 .rowCellEven {
-  background-color: #FFFFFF !important;
+  background-color: #ffffff !important;
 }
 
 .rowCellOdd {
-  background-color: #F4F4F4 !important;
+  background-color: #f4f4f4 !important;
 }
